refactor: 유저 도메인 패키지명 변경
- 도메인 기준에 맞는 네이밍으로 변경했습니다.
